Match Slide Out Control
- Added Slide Out Control
